refactor(mcp): split tools.py into package (H7 — SRP)
Split src/interfaces/mcp/tools.py (1060 lines) into a package:
- tools/_shared.py (134 lines): singletons, init_service, helpers
- tools/memory.py (762 lines): 17 memory tools + serializers
- tools/messaging.py (168 lines): 4 messaging tools + serializers
- tools/__init__.py (85 lines): re-exports for backward compatibility

All 100+ import sites unchanged — __init__.py re-exports everything.
Test patch paths updated: tools._get_* → tools.memory._get_*
1953 tests pass, ruff clean, mypy clean.
